+++ cuda_3d/sph.cuh
@@ -7,15 +7,15 @@
 
 #include "common.h"
 
-__device__ double cubic_kernel(double r_norm, double h) {
-    double k = 8.0 / PI;
+__device__ float cubic_kernel(float r_norm, float h) {
+    float k = 8.0 / PI;
     k = k / (h * h * h);
-    double q = r_norm / h;
+    float q = r_norm / h;
 
     if (q <= 1.0) {
         if (q <= 0.5) {
-            double q2 = q * q;
-            double q3 = q2 * q;
+            float q2 = q * q;
+            float q3 = q2 * q;
             return k * (6.0 * q3 - 6.0 * q2 + 1.0);
         }
         else
@@ -23,27 +23,27 @@ __device__ double cubic_kernel(double r_norm, double h) {
     }
 }
 
-__device__ Vector3d cubic_kernel_derivative(Vector3d& r, double h) {
+__device__ Vector3f cubic_kernel_derivative(Vector3f& r, float h) {
     // Constants
-    double k = 8.0 / PI;
+    float k = 8.0 / PI;
     k = 6.0 * k / (h * h * h);
 
     // Calculate norm of r
-    double r_norm = norm(r);
-    double q = r_norm / h;
+    float r_norm = norm(r);
+    float q = r_norm / h;
 
     // Initialize the result vector
-    Vector3d res = {0.0, 0.0, 0.0};
+    Vector3f res = {0.0, 0.0, 0.0};
 
     if (r_norm > 1e-5 && q <= 1.0) {
-        Vector3d grad_q = {r.x / (r_norm * h), r.y / (r_norm * h), r.z / (r_norm * h)};
+        Vector3f grad_q = {r.x / (r_norm * h), r.y / (r_norm * h), r.z / (r_norm * h)};
 
         if (q <= 0.5) {
             res.x = k * q * (3.0 * q - 2.0) * grad_q.x;
             res.y = k * q * (3.0 * q - 2.0) * grad_q.y;
             res.z = k * q * (3.0 * q - 2.0) * grad_q.z;
         } else {
-            double factor = 1.0 - q;
+            float factor = 1.0 - q;
             res.x = k * (-factor * factor) * grad_q.x;
             res.y = k * (-factor * factor) * grad_q.y;
             res.z = k * (-factor * factor) * grad_q.z;
